seed/requirements.txt
google-cloud-firestore==2.11.1
flask==2.2.5
base58==2.1.0